requirements.txt
pandas~=1.1.1
psycopg2-binary~=2.8.5
